chore: rework remove reference component to use react-query & restyle
