+++ src/js/references/components/Detail/ReferenceSettings.tsx
@@ -5,7 +5,7 @@ import { LoadingPlaceholder, SectionHeader } from "../../../base";
 import { useGetReference } from "../../hooks";
 import { LocalSourceTypes } from "../SourceTypes/LocalSourceTypes";
 import ReferenceMembers from "./ReferenceMembers";
-import RemoveReference from "./Remove";
+import RemoveReference from "./RemoveReference";
 
 type ReferenceSettingsProps = {
     /** Match object containing path information */
@@ -29,11 +29,7 @@ export default function ReferenceSettings({ match }: ReferenceSettingsProps) {
             </SectionHeader>
             <ReferenceMembers noun="user" members={sortBy(data.users, "id")} refId={refId} />
             <ReferenceMembers noun="group" members={sortBy(data.groups, "id")} refId={refId} />
-            <SectionHeader>
-                <h2>Delete</h2>
-                <p>Permanently delete the reference.</p>
-            </SectionHeader>
-            <RemoveReference />
+            <RemoveReference id={data.id} />
         </>
     );
 }
